:art: Refactor SYLink handling to processSiyuanUri and improve URI processing logic (resolve circular dependency)
