+++ app/src/protyle/util/compatibility.ts
@@ -5,7 +5,7 @@ import {Constants} from "../../constants";
 import {ipcRenderer} from "electron";
 /// #endif
 /// #if MOBILE
-import {processSYLink} from "../../editor/openLink";
+import {processSiyuanUri} from "../../util/uri";
 /// #endif
 import {getDefaultSubType, getDefaultType} from "../../search/getDefault";
 import {hideMessage, showMessage} from "../../dialog/message";
@@ -71,7 +71,7 @@ export const openByMobile = (uri: string) => {
         return;
     }
     /// #if MOBILE
-    if (processSYLink(window.siyuan.ws.app, uri)) {
+    if (processSiyuanUri(window.siyuan.ws.app, uri)) {
         return;
     }
     /// #endif
@@ -696,4 +696,3 @@ export const initNativeDialogOverride = () => {
     };
 };
 /// #endif
-
